Remove documentation and unused files, update README with new project focus
