--- README_ENHANCED.md
@@ -1,187 +1,0 @@
-# 🎭 Mickey Mouse Agent - Enhanced Version
-
-A delightful interactive Mickey Mouse agent that can sing, dance, wave, and perform with **actual audio and visual feedback**!
-
-## 🎉 What's New in This Enhanced Version
-
-### 🎵 **Audio Playback**
-- **Real music playback** using Web Audio API in the browser
-- **Simple melodies** generated for each song
-- **Audio feedback** when Mickey sings
-- **Browser-based audio** (no server-side audio dependencies)
-
-### 🎭 **Visual Animations**
-- **Animated Mickey avatar** that responds to actions
-- **CSS animations** for different activities:
-  - 🎤 **Singing**: Mickey scales and rotates while singing
-  - 👋 **Waving**: Mickey rotates side to side
-  - 💃 **Dancing**: Mickey jumps and spins
-  - 😴 **Resting**: Mickey gently pulses and fades
-
-### 🎨 **Enhanced Visual Feedback**
-- **Multiple emoji sequences** for wave styles
-- **Alternating emojis** during singing and dancing
-- **Visual step-by-step feedback** for all actions
-- **Smooth transitions** between different states
-
-## 🚀 Features
-
-### 🎤 **Singing**
-- Choose from 6 classic Disney songs
-- Real-time audio playback in browser
-- Visual feedback with alternating 🎵 and 🎤 emojis
-- Lyrics display with timing
-
-### 👋 **Waving**
-- 5 different wave styles: friendly, excited, royal, magical, goofy
-- Animated wave sequences
-- Visual feedback with style-specific emojis
-
-### 💃 **Dancing**
-- 6 different dance moves
-- Step-by-step visual feedback
-- Alternating 💃 and 🕺 emojis
-- Detailed dance descriptions
-
-### 🎭 **Full Show**
-- Complete performance sequence
-- Automatic transitions between actions
-- Extended visual animations
-- Multiple action sequence
-
-### 😴 **Rest**
-- Energy regeneration system
-- Visual rest sequence
-- Status tracking
-
-## 🛠️ Technical Improvements
-
-### **Frontend Enhancements**
-- **Web Audio API** integration for browser-based audio
-- **CSS animations** with keyframes for smooth transitions
-- **JavaScript audio generation** for simple melodies
-- **Responsive design** improvements
-- **Real-time status updates**
-
-### **Backend Improvements**
-- **Enhanced visual feedback** in console output
-- **Better timing** for visual effects
-- **Improved error handling**
-- **Simplified dependencies** (removed pygame/numpy issues)
-
-### **Audio System**
-- **Browser-based audio generation** using Web Audio API
-- **Simple melody creation** with sine wave oscillators
-- **Automatic audio playback** when actions are triggered
-- **No server-side audio dependencies**
-
-## 🎯 How to Use
-
-### **Running the Application**
-
-1. **Install dependencies:**
-   ```bash
-   python3 -m venv venv
-   source venv/bin/activate
-   pip install -r requirements.txt
-   ```
-
-2. **Start the server:**
-   ```bash
-   python app.py
-   ```
-
-3. **Open in browser:**
-   ```
-   http://localhost:5000
-   ```
-
-### **Testing the API**
-
-Run the test script to verify all functionality:
-```bash
-python test_demo.py
-```
-
-### **API Endpoints**
-
-- `GET /api/status` - Get Mickey's current status
-- `POST /api/sing` - Make Mickey sing a song
-- `POST /api/wave` - Make Mickey wave
-- `POST /api/dance` - Make Mickey dance
-- `POST /api/show` - Make Mickey perform a full show
-- `POST /api/rest` - Make Mickey rest
-
-## 🎨 Visual Features
-
-### **Avatar Animations**
-- **Bounce**: Default idle animation
-- **Sing**: Scale and rotate animation
-- **Wave**: Side-to-side rotation
-- **Dance**: Jump and spin animation
-- **Rest**: Gentle pulse and fade
-
-### **Emoji Sequences**
-- **Wave Styles**: Different emoji combinations for each style
-- **Singing**: Alternating 🎵 and 🎤 emojis
-- **Dancing**: Alternating 💃 and 🕺 emojis
-- **Rest**: Series of 😴 emojis
-
-## 🔧 Technical Details
-
-### **Audio Implementation**
-- Uses **Web Audio API** for browser-based audio
-- **Sine wave oscillators** for simple melodies
-- **Automatic frequency progression** for musical notes
-- **Volume control** and **fade effects**
-
-### **Animation System**
-- **CSS keyframes** for smooth animations
-- **JavaScript class manipulation** for state changes
-- **Transition effects** between different states
-- **Responsive animations** that work on all devices
-
-### **Dependencies**
-- **Flask**: Web framework
-- **Werkzeug**: WSGI utilities
-- **No external audio libraries** (browser-based audio)
-
-## 🐛 Bug Fixes
-
-### **Resolved Issues**
-- ✅ **No audio playback** - Now uses Web Audio API
-- ✅ **No visual feedback** - Added CSS animations and emoji sequences
-- ✅ **Dependency conflicts** - Removed pygame/numpy compatibility issues
-- ✅ **Static emoji display** - Now shows animated sequences
-- ✅ **No browser audio** - Implemented browser-based audio generation
-
-### **Performance Improvements**
-- **Faster loading** without heavy audio dependencies
-- **Smoother animations** with CSS transitions
-- **Better error handling** for audio contexts
-- **Responsive design** improvements
-
-## 🎭 Demo
-
-The enhanced version now provides:
-
-1. **🎵 Real Audio**: Hear Mickey sing with browser-generated melodies
-2. **🎭 Visual Animations**: Watch Mickey animate during actions
-3. **👋 Enhanced Feedback**: See detailed emoji sequences for all actions
-4. **💃 Better Interactions**: More engaging visual and audio experience
-
-## 🌟 Future Enhancements
-
-Potential improvements for future versions:
-- **More complex melodies** with multiple instruments
-- **Sound effects** for different actions
-- **More animation variations**
-- **Voice synthesis** for speaking
-- **Interactive elements** (click to interact)
-- **Mobile app version**
-
----
-
-**🎉 Enjoy the enhanced Mickey Mouse Agent experience!**
-
-*Now with real audio and visual feedback!*
